Changes to mesh naming and exporting
-Added MeshSelector component to choose what mesh an object should use.
- Added icons for scripts.
- Altered direction gizmo drawing to  always show line.
- Made some QOL changes to how Oscillating platform scripts are added: the direction gizmo is automatically created and assigned.
-Renamed meshes so they can be enumed.
- Prefabed objects in the level (springs, oscillators, spikes)
-Moved scripts to new folder.
-Added a tiny example level with one of each type of object. This was made with the intention of new levels being duplicated from this, then expanded upon, such that no parts of the level are missing.
