+++ Assembly-CSharp.csproj
@@ -55,15 +55,16 @@
     <UnityVersion>2022.3.17f1</UnityVersion>
   </PropertyGroup>
   <ItemGroup>
-    <Analyzer Include="C:\Program Files\Microsoft Visual Studio\2022\Community\Common7\IDE\Extensions\Microsoft\Visual Studio Tools for Unity\Analyzers\Microsoft.Unity.Analyzers.dll" />
+    <Analyzer Include="C:\Program Files (x86)\Microsoft Visual Studio\2019\Community\Common7\IDE\Extensions\Microsoft\Visual Studio Tools for Unity\Analyzers\Microsoft.Unity.Analyzers.dll" />
     <Analyzer Include="C:\Program Files\Unity\Hub\Editor\2022.3.17f1\Editor\Data\Tools\Unity.SourceGenerators\Unity.SourceGenerators.dll" />
     <Analyzer Include="C:\Program Files\Unity\Hub\Editor\2022.3.17f1\Editor\Data\Tools\Unity.SourceGenerators\Unity.Properties.SourceGenerator.dll" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Assets\DirectionGizmo.cs" />
+    <Compile Include="Assets\Scripts\DirectionGizmo.cs" />
+    <Compile Include="Assets\Scripts\Spring.cs" />
+    <Compile Include="Assets\Scripts\OscPlat.cs" />
+    <Compile Include="Assets\Scripts\MeshSelector.cs" />
     <Compile Include="Assets\JsonConverter\SceneToJson.cs" />
-    <Compile Include="Assets\OscPlat.cs" />
-    <Compile Include="Assets\Spring.cs" />
   </ItemGroup>
   <ItemGroup>
     <Reference Include="UnityEngine">
