+++ Assets/ExampleLevel.json
@@ -1,0 +1,1 @@
+{"StartPoint":{"x":-91.22000122070313,"y":8.100000381469727,"z":13.5},"EndPoint":{"x":-46.70000076293945,"y":7.099999904632568,"z":13.5},"DeathPlane":{"x":0.0,"y":0.0,"z":0.0},"checkPoints":[{"x":7.289999961853027,"y":3.380000114440918,"z":0.0}],"primitiveGameObject":[{"mesh":"Cube.msh","dimensions":{"x":10.0,"y":0.5,"z":20.0},"rotation":{"x":0.0,"y":0.0,"z":0.0,"w":1.0},"position":{"x":-91.22000122070313,"y":5.099999904632568,"z":15.930000305175782},"inverseMass":1.0,"physicType":"kinematic","volumeType":"box","colliderExtents":{"x":10.0,"y":0.5,"z":20.0},"colliderRadius":0.0,"shouldNetwork":false},{"mesh":"Cube.msh","dimensions":{"x":10.0,"y":0.5,"z":20.0},"rotation":{"x":0.0,"y":0.0,"z":0.0,"w":1.0},"position":{"x":-44.150001525878909,"y":5.099999904632568,"z":15.930000305175782},"inverseMass":1.0,"physicType":"kinematic","volumeType":"box","colliderExtents":{"x":10.0,"y":0.5,"z":20.0},"colliderRadius":0.0,"shouldNetwork":false}],"oscList":[{"mesh":"pipesupphigh.obj","dimensions":{"x":6.0,"y":0.4000000059604645,"z":5.0},"rotation":{"x":0.0,"y":0.0,"z":0.0,"w":1.0},"position":{"x":-58.79999923706055,"y":5.28000020980835,"z":16.59000015258789},"inverseMass":0.0,"physicType":"","volumeType":"box","colliderExtents":{"x":6.0,"y":0.4000000059604645,"z":5.0},"colliderRadius":0.0,"shouldNetwork":true,"timePeriod":5.0,"dist":22.77002716064453,"direction":{"x":-0.9999988079071045,"y":0.0,"z":-0.001537101692520082},"cooldown":0.5,"waitDelay":0.5}],"harmOscList":[{"mesh":"Cube.msh","dimensions":{"x":2.0,"y":2.0,"z":2.0},"rotation":{"x":0.0,"y":0.0,"z":0.0,"w":1.0},"position":{"x":-43.47999954223633,"y":4.28000020980835,"z":22.030000686645509},"inverseMass":0.0,"physicType":"","volumeType":"box","colliderExtents":{"x":2.0,"y":2.0,"z":2.0},"colliderRadius":0.0,"shouldNetwork":true,"timePeriod":0.4000000059604645,"dist":2.0,"direction":{"x":0.0,"y":1.0,"z":0.0},"cooldown":1.0,"waitDelay":1.0}],"pointLights":[{"position":{"x":-64.5999984741211,"y":6.630000114440918,"z":16.219999313354493},"colour":{"x":1.0,"y":1.0,"z":1.0,"w":1.0},"radius":8.050000190734864}],"springs":[{"mesh":"Cube.msh","dimensions":{"x":10.0,"y":1.0,"z":10.0},"rotation":{"x":0.0,"y":0.0,"z":0.0,"w":1.0},"position":{"x":-91.30999755859375,"y":5.179999828338623,"z":27.959999084472658},"inverseMass":0.0,"physicType":"","volumeType":"box","colliderExtents":{"x":10.0,"y":1.0,"z":10.0},"colliderRadius":0.0,"shouldNetwork":false,"direction":{"x":0.8405042886734009,"y":0.47857511043548586,"z":-0.254004567861557},"force":100.0,"activeTime":0.0,"isContinuous":false,"continuousForce":0.0}]}
