+++ Assets/MESHES/updated.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: c31edf68f43f63c4d8a814e77bf60706
-PrefabImporter:
+guid: c7d942e68309ac54786659677d59ba30
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
