+++ Assets/Scenes/Example Level.unity.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: ae403ff7b49274247bc60403e0118924
-PrefabImporter:
+guid: 26b0f8f639ec61945bdb8b26f8940d59
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
