+++ Assets/Scripts/DirectionGizmo.cs.meta
@@ -5,7 +5,7 @@ MonoImporter:
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
-  icon: {instanceID: 0}
+  icon: {fileID: 2800000, guid: a2979d9d5e7fd814382a419f6f6eefae, type: 3}
   userData: 
   assetBundleName: 
   assetBundleVariant:
