+++ Assets/Scripts/Spring.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e19ccfe376cf81347b684df9835a9802
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {fileID: 2800000, guid: f8f96e7b72442cc49a78d58fdeec4ce8, type: 3}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
