+++ Assets/eeeeeee.json.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 34f393eeaeb174246bddb162f68891db
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
