+++ Assets/icons.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a7981faece4052140a50bd489cc39f3d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
