--- Assets/prefabs/HarmfulOscillatingPlatform.prefab
@@ -1,107 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!1001 &8535729988266539983
-PrefabInstance:
-  m_ObjectHideFlags: 0
-  serializedVersion: 2
-  m_Modification:
-    serializedVersion: 3
-    m_TransformParent: {fileID: 0}
-    m_Modifications:
-    - target: {fileID: 56743507979563783, guid: c31edf68f43f63c4d8a814e77bf60706, type: 3}
-      propertyPath: m_Name
-      value: HarmfulOscillatingPlatform
-      objectReference: {fileID: 0}
-    - target: {fileID: 199025994952975368, guid: c31edf68f43f63c4d8a814e77bf60706, type: 3}
-      propertyPath: m_LocalPosition.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 199025994952975368, guid: c31edf68f43f63c4d8a814e77bf60706, type: 3}
-      propertyPath: m_LocalPosition.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 199025994952975368, guid: c31edf68f43f63c4d8a814e77bf60706, type: 3}
-      propertyPath: m_LocalPosition.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 199025994952975368, guid: c31edf68f43f63c4d8a814e77bf60706, type: 3}
-      propertyPath: m_LocalRotation.w
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 199025994952975368, guid: c31edf68f43f63c4d8a814e77bf60706, type: 3}
-      propertyPath: m_LocalRotation.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 199025994952975368, guid: c31edf68f43f63c4d8a814e77bf60706, type: 3}
-      propertyPath: m_LocalRotation.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 199025994952975368, guid: c31edf68f43f63c4d8a814e77bf60706, type: 3}
-      propertyPath: m_LocalRotation.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 199025994952975368, guid: c31edf68f43f63c4d8a814e77bf60706, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 199025994952975368, guid: c31edf68f43f63c4d8a814e77bf60706, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 199025994952975368, guid: c31edf68f43f63c4d8a814e77bf60706, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 6212372627835380448, guid: c31edf68f43f63c4d8a814e77bf60706, type: 3}
-      propertyPath: dist
-      value: 1.1800001
-      objectReference: {fileID: 0}
-    - target: {fileID: 6212372627835380448, guid: c31edf68f43f63c4d8a814e77bf60706, type: 3}
-      propertyPath: cooldown
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 6212372627835380448, guid: c31edf68f43f63c4d8a814e77bf60706, type: 3}
-      propertyPath: position.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 6212372627835380448, guid: c31edf68f43f63c4d8a814e77bf60706, type: 3}
-      propertyPath: position.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 6212372627835380448, guid: c31edf68f43f63c4d8a814e77bf60706, type: 3}
-      propertyPath: position.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 6212372627835380448, guid: c31edf68f43f63c4d8a814e77bf60706, type: 3}
-      propertyPath: direction.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 6212372627835380448, guid: c31edf68f43f63c4d8a814e77bf60706, type: 3}
-      propertyPath: direction.y
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 6212372627835380448, guid: c31edf68f43f63c4d8a814e77bf60706, type: 3}
-      propertyPath: direction.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 6597632079695814134, guid: c31edf68f43f63c4d8a814e77bf60706, type: 3}
-      propertyPath: m_Materials.Array.data[0]
-      value: 
-      objectReference: {fileID: 2100000, guid: bc444849d884e4b49a31ee9c0791b395, type: 2}
-    - target: {fileID: 8344296502881841678, guid: c31edf68f43f63c4d8a814e77bf60706, type: 3}
-      propertyPath: m_LocalPosition.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 8344296502881841678, guid: c31edf68f43f63c4d8a814e77bf60706, type: 3}
-      propertyPath: m_LocalPosition.y
-      value: 2.95
-      objectReference: {fileID: 0}
-    - target: {fileID: 8344296502881841678, guid: c31edf68f43f63c4d8a814e77bf60706, type: 3}
-      propertyPath: m_LocalPosition.z
-      value: 0
-      objectReference: {fileID: 0}
-    m_RemovedComponents: []
-    m_RemovedGameObjects: []
-    m_AddedGameObjects: []
-    m_AddedComponents: []
-  m_SourcePrefab: {fileID: 100100000, guid: c31edf68f43f63c4d8a814e77bf60706, type: 3}
