+++ Assets/prefabs/LevelObjects/Light.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 816028a48371b224396ee047c5058281
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
