+++ Assets/prefabs/LevelObjects/Spike.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: af71c6522d36209478ff36610b5abf89
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
