+++ Assets/prefabs/LevelObjects/Spring.prefab
@@ -1,6 +1,6 @@
 %YAML 1.1
 %TAG !u! tag:unity3d.com,2011:
---- !u!1 &56743507979563783
+--- !u!1 &2873937758581495028
 GameObject:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -8,49 +8,85 @@ GameObject:
   m_PrefabAsset: {fileID: 0}
   serializedVersion: 6
   m_Component:
-  - component: {fileID: 199025994952975368}
-  - component: {fileID: 5622477416262707687}
-  - component: {fileID: 6597632079695814134}
-  - component: {fileID: 4784989217830902381}
-  - component: {fileID: 6212372627835380448}
+  - component: {fileID: 4083079468865713141}
+  - component: {fileID: 8592590487597046677}
+  - component: {fileID: 7386085518728789813}
+  - component: {fileID: 4972839908308560914}
+  - component: {fileID: 47255075432089607}
+  - component: {fileID: 4879927487006942884}
   m_Layer: 0
-  m_Name: OscillatingPlatform
+  m_Name: Spring
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
---- !u!4 &199025994952975368
+--- !u!4 &4083079468865713141
 Transform:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 56743507979563783}
+  m_GameObject: {fileID: 2873937758581495028}
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 18.35, y: 5.28, z: 16.59}
-  m_LocalScale: {x: 6, y: 0.4, z: 5}
+  m_LocalPosition: {x: -91.31, y: 5.18, z: 27.96}
+  m_LocalScale: {x: 10, y: 1, z: 10}
   m_ConstrainProportionsScale: 0
   m_Children:
-  - {fileID: 8344296502881841678}
+  - {fileID: 1697972302295652932}
   m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!33 &5622477416262707687
+--- !u!114 &8592590487597046677
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2873937758581495028}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: e19ccfe376cf81347b684df9835a9802, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  mesh: Cube.msh
+  dimensions: {x: 0, y: 0, z: 0}
+  position: {x: 0, y: 0, z: 0}
+  direction: {x: 0, y: 0, z: 0}
+  force: 100
+  activeTime: 0
+  isContinuous: 0
+  continuousForce: 0
+  directionObject: {fileID: 1697972302295652932}
+--- !u!33 &7386085518728789813
 MeshFilter:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 56743507979563783}
+  m_GameObject: {fileID: 2873937758581495028}
   m_Mesh: {fileID: 10202, guid: 0000000000000000e000000000000000, type: 0}
---- !u!23 &6597632079695814134
+--- !u!114 &4972839908308560914
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2873937758581495028}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 0d9f539d05465bb4ebec3ae0af301c60, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  mesh: 0
+  extension: 1
+--- !u!23 &47255075432089607
 MeshRenderer:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 56743507979563783}
+  m_GameObject: {fileID: 2873937758581495028}
   m_Enabled: 1
   m_CastShadows: 1
   m_ReceiveShadows: 1
@@ -86,13 +122,13 @@ MeshRenderer:
   m_SortingLayer: 0
   m_SortingOrder: 0
   m_AdditionalVertexStreams: {fileID: 0}
---- !u!65 &4784989217830902381
+--- !u!65 &4879927487006942884
 BoxCollider:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 56743507979563783}
+  m_GameObject: {fileID: 2873937758581495028}
   m_Material: {fileID: 0}
   m_IncludeLayers:
     serializedVersion: 2
@@ -107,28 +143,7 @@ BoxCollider:
   serializedVersion: 3
   m_Size: {x: 1, y: 1, z: 1}
   m_Center: {x: 0, y: 0, z: 0}
---- !u!114 &6212372627835380448
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 56743507979563783}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: be7f6d5094c9c4a43a7f9ec0cab01fa6, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  mesh: Cube.msh
-  dimensions: {x: 0.94, y: 0.6, z: 0.12}
-  position: {x: -0.19, y: -0.44, z: -0.43}
-  timePeriod: 5
-  dist: 0.49
-  direction: {x: 0, y: 0, z: 0}
-  cooldown: 0
-  waitDelay: 0
-  directionObj: {fileID: 8211764656799601107}
---- !u!1 &8211764656799601107
+--- !u!1 &7296511869898744681
 GameObject:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -136,41 +151,40 @@ GameObject:
   m_PrefabAsset: {fileID: 0}
   serializedVersion: 6
   m_Component:
-  - component: {fileID: 8344296502881841678}
-  - component: {fileID: 6564674552344613184}
+  - component: {fileID: 1697972302295652932}
+  - component: {fileID: 8839365730428731227}
   m_Layer: 0
-  m_Name: end direction
+  m_Name: DirectionObject
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
---- !u!4 &8344296502881841678
+--- !u!4 &1697972302295652932
 Transform:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 8211764656799601107}
+  m_GameObject: {fileID: 7296511869898744681}
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: -8.7, y: 0, z: -0.007}
+  m_LocalPosition: {x: 0.771, y: 4.39, z: -0.233}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
-  m_Father: {fileID: 199025994952975368}
+  m_Father: {fileID: 4083079468865713141}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &6564674552344613184
+--- !u!114 &8839365730428731227
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 8211764656799601107}
+  m_GameObject: {fileID: 7296511869898744681}
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: dcbdf5e97e6af7849bd233b4417a10af, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  parent: {fileID: 199025994952975368}
-  drawObject: 1
+  parent: {fileID: 4083079468865713141}
