+++ Assets/prefabs/LevelObjects/Spring.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: e2cdf4513adbb6247ac052b5b9762460
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
